HeroSheet-2.20.2022
This is the latest WORKING character sheet.

The only thing this doesn't have working is the Abilities Sub-Skills and the Ability main roll button to include hte progression point system. For now there is a "Prog Tool" skill that calculates your level and that level is now put into the Level field... which the button does work with.

The sub-skills are a different beast entirely.